Improve pass rate for view tests
For this, I've improved the factories used to generate data for the
tests, sometimes using the Faker gem to generate data. A lot of lines
have had to be commented out (considered 'in progress'), which I hope to
resolve by the end of this branch.

I've also ensured the app/services directory gets loaded when the tests
are run, and that SimpleCov is installed and working.